+++ .huskyrc.js
@@ -1,6 +1,6 @@
-function cp(from, to) {
-  return [`cp ${from} ${to}`, `git add ${to}`];
-}
+//function cp(from, to) {
+//  return [`cp ${from} ${to}`, `git add ${to}`];
+//}
 
 function precommit(...workspaces) {
   return workspaces.map((workspace) => `yarn workspace ${workspace} run precommit`);
@@ -9,11 +9,7 @@ function precommit(...workspaces) {
 module.exports = {
   hooks: {
     'pre-commit': [
-      `markdown-source-import "{,!(node_modules)/**/}*.md" --git-add`,
-      ...cp(`README.md`, `source/src/@handbook/babel-plugin/README.md`),
-      ...cp(`README.md`, `source/src/@handbook/source/README.md`),
-      ...cp(`README.md`, `source/src/@handbook/typescript-source-sampler/README.md`),
-      ...cp(`README.md`, `source/src/@handbook/code-block/README.md`),
+      `markdown-source-import "source/src/**/*.md" --git-add`,
       `lint-staged`,
       ...precommit(`source`),
     ].join(' && '),
