+++ package.json
@@ -1,9 +1,9 @@
 {
   "private": true,
   "version": "2020.08.24",
-  "author": "Name <id@mail.com>",
+  "author": "SSen <i@ssen.name>",
   "license": "MIT",
-  "repository": "github:githubid/repository",
+  "repository": "github:rocket-hangar/handbook",
   "workspaces": [
     "source",
     "example",
@@ -16,10 +16,11 @@
     "publish": "yarn run test && yarn run build && yarn workspace source run publish --out-dir $PWD/out/packages",
     "view": "yarn workspace source run view",
     "doctor": "yarn workspace source run doctor",
-    "format": "prettier --write \"**/*.{js,ts,tsx}\"",
-    "lint": "eslint \"**/*.{js,ts,tsx}\"",
     "test": "yarn workspace source run test",
     "coverage": "yarn workspace source run coverage --coverageDirectory $PWD/coverage",
+    "format": "prettier --write \"**/*.{js,ts,tsx}\"",
+    "markdown-source-import": "markdown-source-import source/src/**/*.md",
+    "lint": "eslint \"**/*.{js,ts,tsx}\"",
     "example": "yarn workspace example run build"
   },
   "lint-staged": {
@@ -36,7 +37,6 @@
     "eslint-config-prettier": "^6.11.0",
     "husky": "^4.2.5",
     "lint-staged": "^10.2.13",
-    "markdown-source-import": "^1.3.1",
     "prettier": "^2.1.1",
     "ts-node": "^9.0.0",
     "typescript": "^4.0.2"
